pyproject.toml
[project]
name = "fimbu"
version = "0.2.1"
description = "A thin functionnal layer over litestar and edgy ORM"
authors = [
    {name = "Edimedia Mutoke", email = "kilufyamutoke@gmail.com"},
]
keywords = ["api", "rest", "asgi", "litestar", "starlite", "fimbu", "edgy ORM"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Web Environment",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development",
  "Typing :: Typed",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "Topic :: Internet",
  "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "edgy>=0.10.1",
    "litestar[standard]>=2.8.2",
    "black>=24.4.2",
    "rich-click>=1.8.0",
    "cryptography>=42.0.5",
    "aiosqlite>=0.20.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "argon2-cffi>=23.1.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
shell = [
    "ptpython>=3.0.26",
]
mail = [
    "aiosmtplib>=3.0.1",
    "dnspython>=2.6.1",
    "email-validator>=2.1.1",
    "redis>=5.0.4",
    "blinker>=1.8.1",
]
testing = ["edgy[testing]", "litestar[test]"]
all = [
    "edgy[postgres,mysql,sqlite,testing]",
    "fimbu[mail]",
    "fimbu[shell]",
    "litestar[annotated-types,attrs,brotli,cli,cryptography,jinja,jwt,mako,minijinja,opentelemetry,prometheus,pydantic,redis,sqlalchemy,standard,structlog]",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.scripts]
fimbu = "fimbu.__main__:run_cli"
